% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crime.R
\docType{data}
\name{crime}
\alias{crime}
\title{A Crime dataset is present here}
\format{
A data frame with 86314 rows and 17 variables:
}
\usage{
crime
}
\description{
A dataset including crime data in Huston
}
\keyword{datasets}
