% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab5.R
\name{find_crime_from_month}
\alias{find_crime_from_month}
\title{find_crime_from_month}
\usage{
find_crime_from_month(data, time)
}
\arguments{
\item{data}{crime_data}

\item{time}{month input}
}
\value{
crime_map
}
\description{
find_crime_from_month
}
