% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab5.R
\name{find_crime_from_time_and_type}
\alias{find_crime_from_time_and_type}
\title{find_crime_from_time_and_type}
\usage{
find_crime_from_time_and_type(data, type, time)
}
\arguments{
\item{data}{crime_data}

\item{type}{crime_type}

\item{time}{month input}
}
\value{
crime_map
}
\description{
find_crime_from_time_and_type
}
