% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab5.R
\name{find_crime_from_type}
\alias{find_crime_from_type}
\title{find_crime_from_type}
\usage{
find_crime_from_type(data, type)
}
\arguments{
\item{data}{crim_data}

\item{type}{crime_type}
}
\value{
crime_map
}
\description{
find_crime_from_type
}
