% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab5.R
\name{textInputRow}
\alias{textInputRow}
\title{The Lab5 API package}
\usage{
textInputRow(inputId, label, value = "", ...)
}
\arguments{
\item{inputId}{inputId}

\item{label}{label}

\item{value}{value}

\item{...}{other input}
}
\description{
Using OpenStreetMaps via Stamen Tiles combine with crime and wind data from ggmap.
}
