% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind.R
\docType{data}
\name{wind}
\alias{wind}
\title{A Wind dataset is present here}
\format{
A data frame with 51529 rows and 8 variables:
}
\usage{
wind
}
\description{
A dataset including wind data in Huston
}
\keyword{datasets}
